Rename files and rearrange folders
